requirements.txt
flask
flask_socketio
python-dotenv
eventlet
psutil
gunicorn
